fix dragon body codec + move to registry package
